Refatorado interfaces de transações
-- Criado uma pasta 'interfaces' para manter declarações globais de interfaces e types do projeto
-- Movido todas as interfaces do model de transaction (exceção sendo a interface do documento) para a nova pasta de interfaces globais
-- Adaptado imports das interfaces de trasações pelo código
