+++ src/db/models/transaction.ts
@@ -4,150 +4,8 @@ import { detailsOf } from '../../currencyApi'
 import type User from '../../userApi/user'
 import type { SuportedCurrencies } from '../../currencyApi'
 
-/** Interface base de uma transaction */
-interface Transaction {
-	/** Identificador único dessa transação */
-	txid: string
-	/** Account de destino da transação */
-	account: string
-	/** Amount da transação */
-	amount: number|bigint|string
-	/**
-	 * Status da transação
-	 *
-	 * pending: A transação foi recebida mas ainda não foi confirmada pela rede
-	 *
-	 * confirmed: A transação foi confirmada pela rede e é considerada
-	 * irreversível
-	 */
-	status: 'pending'|'confirmed'
-	/**
-	 * A quantidade de confirmações que uma transação tem. Transações
-	 * confirmadas em um único bloco (como a NANO) não precisam utilizar isso
-	 */
-	confirmations?: number
-	/**
-	 * Timestamp da transação, o servidor principal usa Date, os módulos
-	 * externos transmitem o timestamp em milisegundos como um number
-	 */
-	timestamp: number|Date
-}
-
-/**
- * Interface base para interfaces de comunicação entre o módulo externo
- * e o main server
- */
-interface ExternalModuleTransaction extends Transaction {
-	amount: string
-	timestamp: number
-	/** Identificador da operação da transação no servidor principal */
-	opid: string
-}
-
-/** Interface de uma transação recebida */
-export interface TxReceived extends Transaction {
-	/**
-	 * Identificador da operação da transação no servidor principal; O servidor
-	 * principal irá retornar uma TxReceived com opid caso o módulo externo
-	 * esteja informando uma traqnsação já computada
-	 */
-	opid?: ExternalModuleTransaction['opid']
-	timestamp: ExternalModuleTransaction['timestamp']
-	amount: ExternalModuleTransaction['amount']|number
-}
-
-/** Interface para ordens de envio de transações */
-export interface TxSend {
-	/** Identificador da operação da transação no servidor principal */
-	opid: ExternalModuleTransaction['opid']
-	/** Account de destino da transação */
-	account: ExternalModuleTransaction['account']
-	/** Amount da transação na unidade base da moeda */
-	amount: ExternalModuleTransaction['amount']
-}
-
-/**
- * Interface para atualização de transações recebidas utilizando
- * o evento update_received_tx
- */
-export interface UpdtReceived {
-	/** Identificador da operação da transação no servidor principal */
-	opid: ExternalModuleTransaction['opid']
-	/**
-	 * Status da transação
-	 *
-	 * pending: A transação foi recebida mas ainda não foi confirmada pela rede
-	 *
-	 * confirmed: A transação foi confirmada pela rede e é considerada irreversível
-	 */
-	status: ExternalModuleTransaction['status']
-	/**
-	 * A quantidade de confirmações que uma transação tem. Transações
-	 * confirmadas em um único bloco (como a NANO) não precisam utilizar isso
-	 */
-	confirmations?: ExternalModuleTransaction['confirmations']
-}
-
-/**
- * Interface para atualização de transações enviadas utilizando
- * o evento update_sent_tx
- */
-export interface UpdtSent {
-	/** Identificador da operação da transação no servidor principal */
-	opid: ExternalModuleTransaction['opid']
-	/** Identificador da transação na rede da moeda */
-	txid: ExternalModuleTransaction['txid']
-	/**
-	 * Status da transação
-	 *
-	 * pending: A transação foi recebida mas ainda não foi confirmada pela rede
-	 *
-	 * confirmed: A transação foi confirmada pela rede e é considerada irreversível
-	 */
-	status: ExternalModuleTransaction['status']
-	/**
-	 * A quantidade de confirmações que uma transação tem. Transações
-	 * confirmadas em um único bloco (como a NANO) não precisam utilizar isso
-	 */
-	confirmations?: ExternalModuleTransaction['confirmations']
-	/** O timestamp da transação na rede da moeda */
-	timestamp: ExternalModuleTransaction['timestamp']
-}
-
-export interface CancelledSentTx {
-	/** Identificador da operação da transação que foi cancelada */
-	opid: ExternalModuleTransaction['opid']
-	status: 'cancelled'
-}
-
-/**
- * Interface Transaction do que o servidor principal usa para se comunicar
- * entre seus módulos internos
- */
-export interface TransactionInternal extends Transaction {
-	amount: number|bigint
-	timestamp: Date
-	/** Tipo dessa transação */
-	type: 'receive'|'send'
-}
-
-/**
- * Objeto retornado pelo servidor ao cliente quando é requisitado informações
- * de uma transação
- */
-export interface TxInfo {
-	status: TransactionDoc['status']
-	currency: TransactionDoc['currency']
-	txid: TransactionDoc['txid']
-	account: TransactionDoc['account']
-	amount: number
-	type: TransactionDoc['type']
-	confirmations: TransactionDoc['confirmations']
-	timestamp: TransactionDoc['timestamp']
-}
-
 /** A interface dessa collection */
-interface TransactionDoc extends Document {
+export interface Transaction extends Document {
 	_id: ObjectId
 	/** Referência ao usuário dono dessa transação */
 	userId: User['id']
@@ -163,26 +21,26 @@ interface TransactionDoc extends Document {
 	 * processing: A transação ainda não foi processada no saldo do usuário,
 	 * podendo ser negada quando isso ocorrer (e deletada do db)
 	 */
-	status: ExternalModuleTransaction['status']|'processing'
+	status: 'processing'|'pending'|'confirmed'
 	/** De qual currency essa transação se refere */
 	currency: SuportedCurrencies
 	/** Identificador da transação na rede da moeda */
-	txid: TransactionInternal['txid']
+	txid: string
 	/** Account de destino da transação */
-	account: TransactionInternal['account']
+	account: string
 	/** Amount da transação */
 	amount: Decimal128
 	/** Taxa cobrada para a execução da operação */
 	fee: number
 	/** Tipo dessa transação */
-	type: TransactionInternal['type']
+	type: 'receive'|'send'
 	/**
 	 * A quantidade de confirmações que uma transação tem. Transações
 	 * confirmadas em um único bloco (como a NANO) não precisam utilizar isso
 	 */
-	confirmations?: TransactionInternal['confirmations']
+	confirmations?: number
 	/** O timestamp da transação na rede da moeda */
-	timestamp: TransactionInternal['timestamp']
+	timestamp: Date
 }
 
 /** Schema da collection de transações dos usuários */
@@ -252,7 +110,7 @@ TransactionSchema.index({
 	}
 })
 
-TransactionSchema.pre('validate', function(this: TransactionDoc) {
+TransactionSchema.pre('validate', function(this: Transaction) {
 	if (this.amount instanceof Decimal128)
 		this.amount = this.amount.truncate(detailsOf(this.currency).decimals)
 })
@@ -261,4 +119,4 @@ TransactionSchema.pre('validate', function(this: TransactionDoc) {
  * Model da collection Transactions, responsável por armazenar as transações
  * dos usuários
  */
-export default mongoose.model<TransactionDoc>('Transaction', TransactionSchema)
+export default mongoose.model<Transaction>('Transaction', TransactionSchema)
